user removed becasuse devise integration, cat scaffold generated

+++ db/migrate/20200105232514_create_cats.rb
@@ -2,7 +2,7 @@ class CreateCats < ActiveRecord::Migration[6.0]
   def change
     create_table :cats do |t|
       t.string :name
-      t.numeric :age
+      t.integer :age
 
       t.timestamps
     end
